removed last current compiler warning

+++ SoulFireEngine/src/SoulFire/Platforms/OpenGL/OpenGLVertexArrayObject.cpp
@@ -73,7 +73,7 @@ namespace SoulFire {
 				ShaderDataTypeToOpenGLNativeType(element.m_type),
 				element.m_normalized ? GL_TRUE : GL_FALSE,
 				layout.GetStride(),
-				(const void*)element.m_offset);
+				(const void*)(size_t)element.m_offset);
 			index++;
 		}
